+++ app/src/main/java/io/github/miuzarte/scrcpyforandroid/storage/AppSettings.kt
@@ -90,11 +90,6 @@ class AppSettings(context: Context): Settings(context, "AppSettings") {
     }
 
     companion object {
-        val LANGUAGE_TAG = Pair(
-            stringPreferencesKey("language_tag"),
-            "",
-        )
-
         // Theme
         val THEME_BASE_INDEX = Pair(
             intPreferencesKey("theme_base_index"),
@@ -333,7 +328,6 @@ class AppSettings(context: Context): Settings(context, "AppSettings") {
     @Parcelize
     data class Bundle(
         // Theme
-        val languageTag: String,
         val themeBaseIndex: Int,
         val monet: Boolean,
         val monetSeedIndex: Int,
@@ -403,7 +397,6 @@ class AppSettings(context: Context): Settings(context, "AppSettings") {
 
     private val bundleFields = arrayOf<BundleField<Bundle>>(
         // Theme
-        bundleField(LANGUAGE_TAG) { it.languageTag },
         bundleField(THEME_BASE_INDEX) { it.themeBaseIndex },
         bundleField(MONET) { it.monet },
         bundleField(MONET_SEED_INDEX) { it.monetSeedIndex },
@@ -474,7 +467,6 @@ class AppSettings(context: Context): Settings(context, "AppSettings") {
 
     private fun bundleFromPreferences(preferences: Preferences) = Bundle(
         // Theme
-        languageTag = preferences.read(LANGUAGE_TAG),
         themeBaseIndex = preferences.read(THEME_BASE_INDEX),
         monet = preferences.read(MONET),
         monetSeedIndex = preferences.read(MONET_SEED_INDEX),